+++ lib/Extension/CompletionWorse/CompletionWorseExtension.php
@@ -3,9 +3,14 @@
 namespace Phpactor\Extension\CompletionWorse;
 
 use Closure;
+use Phpactor\Completion\Bridge\TolerantParser\DebugTolerantCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\LimitingCompletor;
-use Phpactor\Completion\Bridge\TolerantParser\ReferenceFinder\NameSearcherCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\ReferenceFinder\ClassLikeCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\ReferenceFinder\ExpressionNameCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\ReferenceFinder\TypeCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\ReferenceFinder\UseNameCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\SourceCodeFilesystem\ScfClassCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\TolerantCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\TypeSuggestionProvider;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\DoctrineAnnotationCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\KeywordCompletor;
@@ -62,6 +67,7 @@ class CompletionWorseExtension implements Extension
     public const NAME_SEARCH_STRATEGY_NONE = 'none';
     public const PARAM_EXPERIMENTAL = 'completion_worse.experimantal';
     public const PARAM_SNIPPETS = 'completion_worse.snippets';
+    public const PARAM_DEBUG = 'completion_worse.debug';
 
     
     public function load(ContainerBuilder $container): void
@@ -84,12 +90,14 @@ class CompletionWorseExtension implements Extension
 
         /** @phpstan-ignore-next-line */
         $defaults['completion_worse.completor.keyword.enabled'] = false;
+        $defaults['completion_worse.completor.constant.enabled'] = false;
 
         $schema->setDefaults(array_merge($defaults, [
             self::PARAM_CLASS_COMPLETOR_LIMIT => 100,
             self::PARAM_NAME_COMPLETION_PRIORITY => self::NAME_SEARCH_STRATEGY_PROXIMITY,
             self::PARAM_SNIPPETS => true,
             self::PARAM_EXPERIMENTAL => false,
+            self::PARAM_DEBUG => false,
         ]));
 
         $descriptions = array_combine(array_map(
@@ -107,6 +115,7 @@ class CompletionWorseExtension implements Extension
 
         /** @phpstan-ignore-next-line */
         $schema->setDescriptions(array_merge($descriptions, [
+            self::PARAM_DEBUG => 'Include debug info in completion results',
             self::PARAM_SNIPPETS => 'Enable or disable completion snippets',
             self::PARAM_EXPERIMENTAL => 'Enable experimental functionality',
             self::PARAM_CLASS_COMPLETOR_LIMIT => 'Suggestion limit for the filesystem based SCF class_completor',
@@ -132,9 +141,18 @@ class CompletionWorseExtension implements Extension
             ]);
         }
 
+        $container->register(TypeSuggestionProvider::class, function (Container $container) {
+            return new TypeSuggestionProvider(
+                $container->get(NameSearcher::class)
+            );
+        });
+
         $container->register(ChainTolerantCompletor::class, function (Container $container) {
             return new ChainTolerantCompletor(
                 array_map(function (string $serviceId) use ($container) {
+                    if ($container->getParameter(self::PARAM_DEBUG)) {
+                        return new DebugTolerantCompletor($container->get($serviceId));
+                    }
                     return $container->get($serviceId);
                 }, $container->get(self::SERVICE_COMPLETOR_MAP)),
                 $container->get('worse_reflection.tolerant_parser')
@@ -303,10 +321,10 @@ class CompletionWorseExtension implements Extension
             'scf_class' => [
                 'Brute force completion for class names (not recommended)',
                 function (Container $container) {
-                    return new LimitingCompletor(new ScfClassCompletor(
+                    return $this->limitCompletor($container, new ScfClassCompletor(
                         $container->get(SourceCodeFilesystemExtension::SERVICE_REGISTRY)->get('composer'),
                         $container->get('class_to_file.file_to_class')
-                    ), $container->getParameter(self::PARAM_CLASS_COMPLETOR_LIMIT));
+                    ));
                 },
             ],
             'local_variable' => [
@@ -328,8 +346,8 @@ class CompletionWorseExtension implements Extension
                     );
                 },
             ],
-            'constant' => [
-                'Completion for constants',
+            'declared_constant' => [
+                'Completion for constants defined in the Phpactor runtime',
                 function (Container $container) {
                     return new WorseConstantCompletor();
                 },
@@ -343,16 +361,42 @@ class CompletionWorseExtension implements Extension
                     );
                 },
             ],
-            'name_search' => [
-                'Completion for class names, constants and functions located in the index',
+            'expression_name_search' => [
+                'Completion for class names, constants and functions at expression positions that are located in the index',
                 function (Container $container) {
-                    return new LimitingCompletor(new NameSearcherCompletor(
+                    return $this->limitCompletor($container, new ExpressionNameCompletor(
                         $container->get(NameSearcher::class),
                         new ObjectFormatter(
                             $container->get(self::SERVICE_COMPLETION_WORSE_SNIPPET_FORMATTERS)
                         ),
                         $container->get(DocumentPrioritizer::class)
-                    ), $container->getParameter(self::PARAM_CLASS_COMPLETOR_LIMIT));
+                    ));
+                },
+            ],
+            'use' => [
+                'Completion for use imports',
+                function (Container $container) {
+                    return $this->limitCompletor($container, new UseNameCompletor(
+                        $container->get(NameSearcher::class),
+                        $container->get(DocumentPrioritizer::class)
+                    ));
+                },
+            ],
+            'class_like' => [
+                'Completion for class like contexts',
+                function (Container $container) {
+                    return $this->limitCompletor($container, new ClassLikeCompletor(
+                        $container->get(NameSearcher::class),
+                        $container->get(DocumentPrioritizer::class)
+                    ));
+                },
+            ],
+            'type' => [
+                'Completion for types',
+                function (Container $container) {
+                    return $this->limitCompletor($container, new TypeCompletor(
+                        $container->get(TypeSuggestionProvider::class)
+                    ));
                 },
             ],
             'keyword' => [
@@ -365,7 +409,7 @@ class CompletionWorseExtension implements Extension
                 'Docblock completion',
                 function (Container $container) {
                     return new DocblockCompletor(
-                        new TypeSuggestionProvider($container->get(NameSearcher::class)),
+                        $container->get(TypeSuggestionProvider::class),
                         $container->get(WorseReflectionExtension::SERVICE_PARSER)
                     );
                 },
@@ -387,5 +431,15 @@ class CompletionWorseExtension implements Extension
     private function completorEnabledKey(string $key): string
     {
         return sprintf('completion_worse.completor.%s.enabled', $key);
+    }
+
+    private function limitCompletor(Container $container, TolerantCompletor $completor): TolerantCompletor
+    {
+        $limit = $container->getParameter(self::PARAM_CLASS_COMPLETOR_LIMIT);
+        if (null === $limit) {
+            return $completor;
+        }
+
+        return new LimitingCompletor($completor, $limit);
     }
 }