+++ lib/Indexer/Adapter/ReferenceFinder/IndexedNameSearcher.php
@@ -8,6 +8,7 @@ use Phpactor\Indexer\Model\Record\HasPath;
 use Phpactor\Indexer\Model\SearchClient;
 use Phpactor\Name\FullyQualifiedName;
 use Phpactor\ReferenceFinder\NameSearcher;
+use Phpactor\ReferenceFinder\NameSearcherType;
 use Phpactor\ReferenceFinder\Search\NameSearchResult;
 use Phpactor\TextDocument\TextDocumentUri;
 
@@ -20,14 +21,56 @@ class IndexedNameSearcher implements NameSearcher
         $this->client = $client;
     }
 
-    public function search(string $name): Generator
+    /**
+     * @param null|NameSearcherType::* $type
+     */
+    public function search(string $name, ?string $type = null): Generator
     {
-        foreach ($this->client->search(Criteria::shortNameBeginsWith($name)) as $result) {
+        $criteria = Criteria::shortNameBeginsWith($name);
+
+        $typeCriteria = $this->resolveTypeCriteria($type);
+
+        if ($typeCriteria) {
+            $criteria = Criteria::and(
+                $criteria,
+                Criteria::or(
+                    $typeCriteria,
+
+                    // B/C for old indexes
+                    Criteria::isClassTypeUndefined()
+                )
+            );
+        }
+
+        foreach ($this->client->search($criteria) as $result) {
             yield NameSearchResult::create(
                 $result->recordType(),
                 FullyQualifiedName::fromString($result->identifier()),
                 $result instanceof HasPath ? TextDocumentUri::fromString($result->filepath()) : null
             );
         }
+    }
+
+    /**
+     * @param null|NameSearcherType::* $type
+     */
+    private function resolveTypeCriteria(?string $type): ?Criteria
+    {
+        if ($type === NameSearcherType::CLASS_) {
+            return Criteria::isClassConcrete();
+        }
+
+        if ($type === NameSearcherType::INTERFACE) {
+            return Criteria::isClassInterface();
+        }
+
+        if ($type === NameSearcherType::TRAIT) {
+            return Criteria::isClassTrait();
+        }
+        if ($type === NameSearcherType::ENUM) {
+            return Criteria::isClassEnum();
+        }
+
+        return null;
     }
 }