+++ lib/Indexer/Adapter/Tolerant/Indexer/EnumDeclarationIndexer.php
@@ -4,6 +4,7 @@ namespace Phpactor\Indexer\Adapter\Tolerant\Indexer;
 
 use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\Node\Statement\EnumDeclaration;
+use Phpactor\Indexer\Model\Record\ClassRecord;
 use Phpactor\TextDocument\TextDocument;
 use Phpactor\Indexer\Model\Index;
 
@@ -17,7 +18,7 @@ class EnumDeclarationIndexer extends AbstractClassLikeIndexer
     public function index(Index $index, TextDocument $document, Node $node): void
     {
         assert($node instanceof EnumDeclaration);
-        $record = $this->getClassLikeRecord('class', $node, $index, $document);
+        $record = $this->getClassLikeRecord(ClassRecord::TYPE_ENUM, $node, $index, $document);
 
         $this->removeImplementations($index, $record);
         $record->clearImplemented();