+++ lib/Indexer/Adapter/Tolerant/Indexer/TraitDeclarationIndexer.php
@@ -5,6 +5,7 @@ namespace Phpactor\Indexer\Adapter\Tolerant\Indexer;
 use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\Node\Statement\TraitDeclaration;
 use Phpactor\Indexer\Model\Index;
+use Phpactor\Indexer\Model\Record\ClassRecord;
 use Phpactor\TextDocument\TextDocument;
 
 class TraitDeclarationIndexer extends AbstractClassLikeIndexer
@@ -17,7 +18,7 @@ class TraitDeclarationIndexer extends AbstractClassLikeIndexer
     public function index(Index $index, TextDocument $document, Node $node): void
     {
         assert($node instanceof TraitDeclaration);
-        $record = $this->getClassLikeRecord('trait', $node, $index, $document);
+        $record = $this->getClassLikeRecord(ClassRecord::TYPE_TRAIT, $node, $index, $document);
         $index->write($record);
     }
 }