+++ lib/Indexer/Model/Record/ClassRecord.php
@@ -11,6 +11,10 @@ final class ClassRecord implements Record, HasFileReferences, HasPath, HasFullyQ
     use HasFileReferencesTrait;
     use HasPathTrait;
     public const RECORD_TYPE = 'class';
+    public const TYPE_CLASS = 'class';
+    public const TYPE_INTERFACE = 'interface';
+    public const TYPE_TRAIT = 'trait';
+    public const TYPE_ENUM = 'enum';
 
     /**
      * @var array<string>
@@ -25,7 +29,7 @@ final class ClassRecord implements Record, HasFileReferences, HasPath, HasFullyQ
     /**
      * Type of "class": class, interface or trait
      */
-    private string $type;
+    private ?string $type = null;
 
     public static function fromName(string $name): self
     {
@@ -98,5 +102,12 @@ final class ClassRecord implements Record, HasFileReferences, HasPath, HasFullyQ
     public function recordType(): string
     {
         return self::RECORD_TYPE;
+    }
+
+    public function withType(string $type): ClassRecord
+    {
+        $clone = clone $this;
+        $clone->type = $type;
+        return $clone;
     }
 }