+++ lib/ReferenceFinder/NameSearcher.php
@@ -9,6 +9,7 @@ interface NameSearcher
 {
     /**
      * @return Generator<NameSearchResult>
+     * @param NameSearcherType::* $type
      */
-    public function search(string $search): Generator;
+    public function search(string $search, ?string $type = null): Generator;
 }