+++ lib/ReferenceFinder/NameSearcherType.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Phpactor\ReferenceFinder;
+
+class NameSearcherType
+{
+    const FUNCTION = 'function';
+    const CLASS_ = 'class';
+    const INTERFACE = 'interface';
+    const TRAIT = 'trait';
+    const ENUM = 'enum';
+}