+++ lib/ReferenceFinder/Search/NullNameSearcher.php
@@ -7,7 +7,7 @@ use Phpactor\ReferenceFinder\NameSearcher;
 
 class NullNameSearcher implements NameSearcher
 {
-    public function search(string $search): Generator
+    public function search(string $search, ?string $type = null): Generator
     {
         yield from [];
     }